seperated chess notatchess notation into two classes and moved assets to public folder
