+++ public/index.html
@@ -23,6 +23,7 @@
 	<script src="piece.js" type="module"></script>
 	<script src="board.js" type="module"></script>
 	<script src="chess_notation.js" type="module"></script>
+	<script src="fen_notation.js" type="module"></script>
 	<script src="gui.js" type="module"></script>
 
 	<script src="init.js" type="module"></script>
