+++ src/init.ts
@@ -1,8 +1,9 @@
-import { ChessNotation } from "./chess_notation.js";
+import { FEN } from "./fen_notation.js";
 import { GUI } from "./gui.js";
 
 window.onload = function () {
-	const board = ChessNotation.boardFromFEN("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1");
+	const fen = new FEN("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1");
+	const board = fen.load();
 
 	const gui = new GUI(board);
 	console.log(gui);
